New python modules for writing received data into a logfile.
AbstractFileWriter.py provides basic writing functionality.
SimpleFileWriter.py is derived from that class and will be the used
class in the final project. For now it does not implement own
methods. Thos will be added later on.

Furthermore the old data server file was moved DataServer.py and the
actual invocation of the server is now done in main.py.